+++ src/index.ts
@@ -63,6 +63,8 @@ export { routerPlugin } from "./plugins/router/plugin";
 
 export * as logs from "./plugins/logs/api";
 export { logsPlugin } from "./plugins/logs/plugin";
+export * as e2e from "./plugins/e2e/api";
+export { e2ePlugin } from "./plugins/e2e/plugin";
 export * as seo from "./plugins/seo/api";
 export * as server from "./plugins/server/api";
 export { serverPlugin } from "./plugins/server/plugin";